Use uint64_t for script memory buffer
Add <cstdint> and switch script_memory_buffer APIs from size_t/int to uint64_t for size and index parameters/returns in both header and implementation. Update null check to explicit nullptr comparison and adjust make/get_element_size signatures and one-line at() overload accordingly. These changes standardize integer widths for large buffers and ensure consistent cross-translation-unit types.

+++ src/misc_functions.h
@@ -11,6 +11,7 @@
 */
 
 #pragma once
+#include <cstdint>
 #include <string>
 #include <angelscript.h>
 #include "nvgt.h"
@@ -39,16 +40,16 @@ public:
 };
 struct script_memory_buffer {
 	void* ptr;
-	size_t size;
+	uint64_t size;
 	asITypeInfo* subtype;
 	int subtypeid;
-	script_memory_buffer(asITypeInfo* subtype, void* ptr, size_t size) : ptr(ptr), size(size), subtype(subtype), subtypeid(subtype->GetSubTypeId()) {}
-	const void* at(size_t index) const;
-	void* at(size_t index);
+	script_memory_buffer(asITypeInfo* subtype, void* ptr, uint64_t size) : ptr(ptr), size(size), subtype(subtype), subtypeid(subtype->GetSubTypeId()) {}
+	const void* at(uint64_t index) const;
+	void* at(uint64_t index);
 	CScriptArray* to_array() const;
 	script_memory_buffer& from_array(CScriptArray* array);
-	int get_element_size() const;
-	static void make(script_memory_buffer* mem, asITypeInfo* subtype, void* ptr, int size);
+	uint64_t get_element_size() const;
+	static void make(script_memory_buffer* mem, asITypeInfo* subtype, void* ptr, uint64_t size);
 	static void copy(script_memory_buffer* mem, asITypeInfo* subtype, const script_memory_buffer& other);
 	static void destroy(script_memory_buffer* mem);
 	static bool verify(asITypeInfo *subtype, bool& no_gc);
